+++ course-work/implementations/JobListing/src/main/java/com/example/JobListing/Infrastructure/ResponseDTOs/JobListingTechResponseDTO.java
@@ -5,9 +5,9 @@ import lombok.Builder;
 @Builder
 public record JobListingTechResponseDTO
     (
-        int id,
-        int listing_id,
-        int tech_id,
+        Long id,
+        Long listing_id,
+        Long tech_id,
         boolean IsRequired
     )
 { }
