+++ course-work/implementations/JobListing/src/main/java/com/example/JobListing/Repository/IBaseRepository.java
@@ -4,4 +4,4 @@ import com.example.JobListing.Entities.BaseEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IBaseRepository<T extends BaseEntity> extends JpaRepository<T, Integer> { }
+public interface IBaseRepository<T extends BaseEntity> extends JpaRepository<T, Long> { }
